+++ Changelog.md
@@ -12,6 +12,7 @@
 - Add support for Delete user by username
 - Add support for `*` in IP restrictions
 - Fix user role `None` when empty role in Auth Codes
+- Add Auth code on Authentication endpoint
 
 ## 2.6.2 (29 April 2021)
 - Update documentation link with the plugin website URL